moved all sorting functionality in productRoute.js to helper.js for more organized, cleaner, code
